% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_time.R
\name{add_local_time}
\alias{add_local_time}
\title{Add local time from UTC for US counties}
\usage{
add_local_time(df, date_colname)
}
\arguments{
\item{df}{A dataframe with columns with county FIPS numbers (\code{fips})
and datetime in UTC.}

\item{date_colname}{A character string giving the name of the column in
the dataframe with a datetime in the UTC timezone.}
}
\value{
This function returns the dataframe that was input, but with added
   columns for \code{local_time} (a character string giving the local date
   in the format "\%Y\%m\%d\%H\%m\%s") and \code{local_date} (a Date object
   giving the date in the local time).
}
\description{
Add local time from UTC for US counties
}
\note{
The local time is given as a character
   string, rather than a POSIXct object, so that it can have different time
   zones for different counties within the same dataframe (i.e., if there
   are two counties in the dataframe that are in different time zones).
}
\examples{
library(hurricaneexposure)
data(closest_dist)
library(lubridate)
floyd <- filter(closest_dist, storm_id == "Floyd-1999") \%>\%
         mutate(closest_date = ymd_hm(closest_date, tz = "UTC"))
floyd <- add_local_time(df = floyd, date_colname = "closest_date")

}

