% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_time.R
\name{calc_local_time}
\alias{calc_local_time}
\title{Calculate local time from UTC for US counties}
\usage{
calc_local_time(date_time, fips, include_tz = TRUE)
}
\arguments{
\item{date_time}{The vector of the date-time of each observation in
Coordinated Universal Time (UTC). This vector can either have a
\code{POSIXct} class or be a character string, with date-time given
as four-digit year, two-digit month, two-digit day, two-digit hour,
and two-digit minutes (with hours based on a 24-hour system). Examples
of acceptable formats include, for the example of 1:00 PM Jan. 2 1999,
"199901021300", "1999-01-02 13:00", and "1999/01/02 13:00".}

\item{fips}{A character vector giving the 5-digit FIPS code of the county
associated with each observation. This can be either a string of length 1
or a vector as long as the \code{date_time} vector, if different
observations come from different counties.}

\item{include_tz}{A TRUE / FALSE value specifying whether to include a
column with the local time zone (\code{local_tz}) in the final output.}
}
\value{
This function returns the dataframe that was input, but with added
   columns for \code{local_time} (a character string giving the local date
   in the format "\%Y\%m\%d\%H\%m\%s") and \code{local_date} (a Date object
   giving the date in the local time).
}
\description{
This function can input date-time values in Coordinated Universal Time (UTC;
also known as Zulu Time), along with a vector with county Federal Information
Processing Standard (FIPS) codes, and calculates the local date-time as
well as the local date based on th UTC date-time.
}
\details{
Because this function can calculate local times for a dataframe
that includes counties in different time zones, it outputs the local
date-time as a character vector, rather than a date-time object, because
it seems that a vector of date-time class can only have one associated
time zone.
}
\note{
The local time is given as a character
   string, rather than a POSIXct object, so that it can have different time
   zones for different counties within the same dataframe (i.e., if there
   are two counties in the dataframe that are in different time zones).
}
\examples{
ex_datetime <- "1999-01-01 08:00"
ex_fips <- "36061"
calc_local_time(date_time = ex_datetime, fips = ex_fips)

ex_datetime <- c("1999-01-01 08:00", "1999-01-01 09:00",
                 "1999-01-01 10:00")
ex_fips <- "36061"
calc_local_time(date_time = ex_datetime, fips = ex_fips)

ex_datetime <- c("1999-01-01 08:00", "1999-01-01 09:00",
                 "1999-01-01 10:00")
ex_fips <- c("36061", "17031", "06037")
calc_local_time(date_time = ex_datetime, fips = ex_fips)
}

