% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_local_time.R
\name{calc_single_datetime}
\alias{calc_single_datetime}
\title{Calculate local time for a single observation}
\usage{
calc_single_datetime(datetime, tz)
}
\arguments{
\item{datetime}{A POSIXct object in UTC timezone}

\item{tz}{A character string giving the local timezone}
}
\value{
A character string giving the time in the local timezone
}
\description{
Calculate local time for a single observation
}
\note{
This must output the date as a character string, because otherwise
   all dates will be transformed to numeric values when you run the
   function through and \code{apply} function.
}
\examples{
utc_time <- as.POSIXct("1999-09-15 14:30:00", tz = "GMT")
local_time <- calc_single_datetime(utc_time, tz = "US/Eastern")

}

