% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_tzs}
\alias{county_tzs}
\title{County time zone designations}
\format{An object of class \code{data.frame} with 3143 rows and 4 columns.}
\usage{
county_tzs
}
\description{
County time zone designations
}
\keyword{datasets}

